% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate}
\alias{calculate}
\title{Calculate derived data}
\usage{
calculate(data, data_type, ..., value_fun, error_fun = function(...)
  return(NA), name_fun = function(...) return(paste(list(...), collapse =
  " ")), filter_new = NULL, quiet = F)
}
\arguments{
\item{data}{a data frame with raw ion counts retrieved from \code{\link{load_analysis_data()}}, can be grouped to do calculations within individual groups}

\item{...}{the columns to send to the value, error and naming function for each derived value, e.g. c("12C", "13C", "12C14N"), the number of parameters needs to match those expected by the value, error and name functions. Error values of different columns (say for classical error propagation) can be addressed using the suffix "sigma", e.g. c("12C", "12C sigma") would pass both the value and error of this variable to the value and error function.}

\item{value_fun}{a custom function used to calculate the derived value - needs to match the sets of paramters provided through ...}

\item{error_fun}{a custom function used to calcluate the error (sigma) for the derived value}

\item{name_fun}{a custom function used to construct the variable name for the derived quantity}

\item{filter_new}{an expression to apply as a filter on the new data rows (e.g. plane == "all")}

\item{quiet}{whether the function should output information messages or be quiet (default is to output)}
}
\value{
the original data frame with the sums information appended (data_type == "ion_sum")
}
\description{
This function allows easy calculation of any quantities derived from other variables. The new quantities can be assigned to a specific data_type and values, errors as well as the resulting variable names are calculated/constructed based on custom functions that can be provided via the function parameters. \link{calculate_sums}, \link{calculate_ratios} and \link{calculate abundances} are all based on this and provide an easy way for common standard calculations.
}

