% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{extract_roi_boundaries}
\alias{extract_roi_boundaries}
\title{figure out the ROI boundaries from LANS map export}
\usage{
extract_roi_boundaries(data)
}
\arguments{
\item{data}{data_frame from LANS map, can already be grouped}
}
\description{
figure out the ROI boundaries from LANS map export
}

