% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{lans2r-package}
\alias{lans2r}
\alias{lans2r-package}
\title{lans2r: Work with Look at NanoSIMS Data in R}
\description{
R interface for working with nanometer scale secondary ion mass
spectrometry (NanoSIMS) data exported from Look at NanoSIMS.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KopfLab/lans2r}
  \item Report bugs at \url{https://github.com/KopfLab/lans2r/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Kopf \email{sebastian.kopf@colorado.edu}

}
\keyword{internal}
