% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{lans2r}
\alias{lans2r}
\alias{lans2r-package}
\title{lans2r package}
\description{
R interface for working with NanoSIMS ROI data defined in Look at NanoSIMS

To get started, take alook at the tutorial R markdown file.
}
\author{
Sebastian Kopf
}

