% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hmr_files.R
\name{load_HMR}
\alias{load_HMR}
\title{Read data from HMR txt files}
\usage{
load_HMR(folder, prefix, suffix = ".hmr_txt")
}
\arguments{
\item{folder}{\itemize{
\item the folder where the HMR files are
}}

\item{prefix}{\itemize{
\item the prefix for the set of hmr files
}}

\item{suffix}{\itemize{
\item the suffix for the set of hmr files
}}
}
\description{
See HMR vignette for examples.
}
