% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_maps}
\alias{plot_maps}
\title{Plot NanoSIMS ion maps}
\usage{
plot_maps(
  data,
  draw_ROIs = TRUE,
  normalize = TRUE,
  color_scale = c("black", "white")
)
}
\arguments{
\item{data}{the ion maps data frame}

\item{draw_ROIs}{whether to draw in the ROIs or not (default TRUE)}

\item{normalize}{whether to normalize the intensity scale for each panel (default TRUE)}

\item{color_scale}{what color scale to use for high and low intensity, default is black & white}
}
\description{
Helps to plot the ion maps exported and loaded from LANS. Can overlay the ROI boundaries
for clarity as well. Note that this does not currently support any smoothing yet so plotting
ratios or abundances will most likely not work well because individual pixels have extreme
values.
}
