% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{spread_data}
\alias{spread_data}
\title{Spread data into wide format}
\usage{
spread_data(data, values = TRUE, errors = TRUE)
}
\arguments{
\item{data}{a data frame with lans2r data}

\item{values}{whether to include the values in wide format}

\item{errors}{whether to include the errors in wide format}
}
\value{
the original data frame but in wide format
}
\description{
This function allows easy spreading into wide format.
}
